Segment some more
