+++ src/apps/admin/adminportal/AdminPortal/Page/ViewScopes.svelte
@@ -47,7 +47,7 @@
   async function save() {
     if (
       scopeList.includes(AdminScopes.adminGod) &&
-      !(await process.userDaemon!.Confirm(
+      !(await process.userDaemon!.helpers!.Confirm(
         "Are you sure?",
         "This user has the <code>admin.god</code> scope applied. This means that they can access all administrative functions of Sacruda, regardless of any other scopes. Are you sure you want to save these changes?",
         "Cancel",
