+++ src/apps/admin/adminportal/AdminPortal/Page/ViewShare/ChangeQuota.svelte
@@ -12,7 +12,7 @@
 
   async function changeQuota() {
     if (!newQuota) return;
-    const confirm = await process.userDaemon?.Confirm(
+    const confirm = await process.userDaemon?.helpers!.Confirm(
       "Confirm quota change?",
       `Are you sure you want to change the quota of '${share.shareName}' to ${formatBytes(newQuota)}?`,
       "Cancel",
