+++ src/apps/admin/adminportal/AdminPortal/Page/ViewStoreItem.svelte
@@ -194,7 +194,7 @@
 
   async function toggleOfficial() {
     if (item.official) {
-      const confirm = await process.userDaemon?.Confirm(
+      const confirm = await process.userDaemon?.helpers!.Confirm(
         "Mark store item?",
         "Are you sure you want to mark this store item as NOT official? This implies that the package was not made by ArcOS administrators and does NOT act as an official utility or optional feature for ArcOS",
         "Cancel",
@@ -207,7 +207,7 @@
         await process.switchPage("viewStoreItem", { id: item._id }, true);
       }
     } else {
-      const confirm = await process.userDaemon?.Confirm(
+      const confirm = await process.userDaemon?.helpers!.Confirm(
         "Mark store item?",
         "Are you sure you want to mark this store item as official? This implies that the package was made by ArcOS administrators and acts as an official utility or optional feature for ArcOS",
         "Cancel",
@@ -223,7 +223,7 @@
   }
 
   async function deleteItem() {
-    const confirm = await process.userDaemon?.Confirm(
+    const confirm = await process.userDaemon?.helpers!.Confirm(
       "Delete store item?",
       "Are you sure you want to delete this store item? This should only ever happen if the user has lost access to their account, or if the package conflicts with another package.",
       "Cancel",
