+++ src/apps/admin/adminportal/AdminPortal/Page/ViewUser/ChangeEmail.svelte
@@ -7,7 +7,7 @@
   let loading = $state<boolean>(false);
 
   async function rename() {
-    const confirm = await process.userDaemon?.Confirm(
+    const confirm = await process.userDaemon?.helpers!.Confirm(
       "Confirm rename?",
       `Are you sure you want to change the email of '${user.username}' to '${newEmail}'?`,
       "Cancel",
