+++ src/apps/components/sharecreategui/runtime.ts
@@ -58,7 +58,7 @@ export class ShareCreateGuiRuntime extends AppProcess {
     const path = `${drive.uuid}:/`;
     const parent = KernelStack().getProcess(this.parentPid);
 
-    if (parent && this.userDaemon?.ParentIs(this, "fileManager")) {
+    if (parent && this.userDaemon?.helpers!.ParentIs(this, "fileManager")) {
       // In case the parent is a file manager, navigate it instead
       const dispatch = KernelStack().ConnectDispatch(this.parentPid);
       dispatch?.dispatch("navigate", path);
